core: Limit "User is away" to 1hr or when changed
Increase the "* User is away" silencing to one hour, and add tracking
for changes in away state.  Only print the "* User is away" if away
state has changed since last away message (including removing and
applying the same away message), or if 1 hour has elapsed.

This reduces the spamming of "* User is away" when one often knows
that, yes, the user is away.  Any change in away state or message
will still be shown.