client: Don't store the list of properties in every PropertyMapItem
It's rather inefficient to store a QStringList in every item, since
it is the same for every type of item anyway. Use a pure virtual
accessor in the base class instead, and return a static QStringList
in every item type.

This prevents the usecase of adding properties dynamically after
creation, but we didn't use that feature anyway, so remove it.