tests: Add unit test for basic SignalProxy functionality
Add unit test cases for attached signals as well as syncable objects.
These test cases don't check for error behavior (yet); they mostly
are intended for ensuring that the upcoming refactoring of the
SignalProxy does not affect the format on the wire. As such, the test
cases mostly focus on the protocol messages sent by invoking
SignalProxy and SyncableObject functionality.