sigproxy: Actually rename SyncableObjects when requested
Until now, renaming a SyncableObject would properly update the maps
in SignalProxy, but it would not actually update the objectName
property on the client side. Fix this.

Remove the renameObject() method in SyncableObject in favor of
using setObjectName() and listening to the notifier signal emitted
by Qt for triggering the update. This ensures that the actual object
name and the expected one stay in sync. Also change places where
we unnecessarily rename right after construction to use the
appropriate constructor instead.

Extend the test case to cover the renaming of SyncableObject.