modernize: Replace init setters/getters by properties
Init setters and getters are one of the more obscure features of
SyncableObject. In addition to reading and syncing the state of all
properties declared with the Q_PROPERTY macro when initializing a
client-side SyncableObject, any slot with a name starting with "init"
is be called, and its return value added to the property map in
the InitData message. The client then calls the corresponding initSet
slot with that value.

This feature was intended to allow for initialization more complex
than just setting property values. However, since the accessors for
a property can be freely defined, and because a property does not
actually need a corresponding data member (a fact which we probably
didn't realize when we originally designed init setters/getters),
this behavior can just as well be achieved by declaring a property
and its potentially complex READ and WRITE functions accordingly.

Using properties directly is also more efficient than going through
the list of all slots, mangling method names, manually marshalling
arguments, and invoking the slots by name; the moc generates more
efficient code for calling property accessors.

Because the format of the InitData message luckily does not actually
distinguish between normal properties and init getter return values,
we can easily and transparently replace init getters/setters by
declaring a property named like the original method base name,
without breaking protocol. The only caveat is that the method base
name starts with a capital letter, slightly violating our naming
conventions for property names, but that's a small price to pay
for getting rid of the init setter/getter feature altogether.

Do this for all init getters/setters, rename the methods as
appropriate, un-slot them to avoid unnecessary moc-generated code,
and clean up some things while we're at it.