syncobj: Automatically sync properties that have a NOTIFY signal
Qt allows to declare a NOTIFY signal for properties. Use this
feature to automatically sync a SyncableObject's properties that
declare such a signal, avoiding the need to manually call SYNC() in
the setter.

NOTIFY signals need to adhere to the following naming scheme, so we
can derive the slot name to be called on receiving a SyncMessage:

- Signals "myFooChanged()" or "myFooSet()" will invoke "setMyFoo()"
- Signal "syncWhateverSlot()" will invoke "whateverSlot()", allowing
  for (almost) arbitrary slot names (they need to start with a
  lower-case letter)

NOTIFY signals can have arbitrary (or no) arguments, because the
property's value is read explicitly when handling the signal.

Modify and extend SignalProxyTest to cover this use case.