syncobj: Directly sync properties
Following up to the automatic sending of property updates for
properties with a supported NOTIFY signal, use the same heuristics
to directly update the property from the SyncMessage on the client
side. This means that instead of invoking the setter dynamically,
QMetaProperty::write() is called, which is more efficient and pushes
the unmarshalling of the value to Qt.

The same heuristics are used for deriving the setter name from the
NOTIFY signal as on the client side as on the sending side, so this
works for all automatically synced properties as long as they are
writable.

If no matching property is found, or the value could not be set,
the slot will be invoked the old-fashioned way.