travis: Add Ubuntu 18.10 "Cosmic" to the list of CI distros
Add Ubuntu 18.10 to the build matrix, so we ensure we can build
on the newest Ubuntu release.

Rework the combination of build options; it's probably enough to
test building without KDE only on the baseline, since the non-KDE
code does not change between distro versions (while the installed
version of KDE Frameworks does). This keeps the number of build
jobs reasonably low.