syncobj: Modernize the calling of sync methods
Extend the SYNCABLE_OBJECT macro to allow for explicitly specifying
syncable methods. Use this information, modern C++ features and some
template meta programming magic to provide the means for handling
SyncMessages and calling sync methods without having to rely on
inefficent runtime look-up of methods, the deprecated Qt slot feature,
or undocumented Qt APIs.

For every sync method declared in the SYNCABLE_OBJECT macro, generate
a call wrapper that can be directly invoked by name, and handles the
forwarding of marshalled arguments as well as request/receive method
functionality. Let SignalProxy make use of this feature for sync
methods that have been declared accordingly, while falling back to
the legacy way if necessary. Adapt SignalProxyTest so it uses and
tests the new implementation.

Adapt the existing SyncableObject specializations to using the new
SYNCABLE_OBJECT macro syntax, albeit without declaring their sync
methods for now; this will be added in follow-up commits.

This brings us one step further in our quest to modernize SignalProxy
and SyncableObject, and paves another part of the way for the removal
of the legacy, inefficent, and hackish way we've been using so far
for mapping protocol messages to function calls.